Create README.md
This project holds a two cpp files and one header file.  The first cpp file is Time class which has a function that alter military time into proper time.  The header file of Time has the prototype Time.  The TimeDriver cpp file is to test the Time classes using Objects.